+++ changelog.md
@@ -5,7 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v11.4.0...HEAD)
 ### Fixed
-* [STREAM-1028](https://inindca.atlassian.net/browse/STREAM-1025) - Fix issue preventing monitoring observers from ending `liveScreenMonitoring` sessions.
+* [STREAM-1028](https://inindca.atlassian.net/browse/STREAM-1025) - Fix issue preventing monitoring observers from ending `liveScreenMonitoring` sessions and add accept session logic for observers.
 * [STREAM-1027](https://inindca.atlassian.net/browse/STREAM-992) - Fix jid to session type evaluation for live monitor jids
 
 # [v11.4.0](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v11.3.4...v11.4.0)